+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Upcoming
+
+## New Component
+
+- Add Chip component
+
 # 6.10.0
 
 ## New Component